[IMP] mail, mail_bot: move notif request and OdooBot name into mail
task-2282334
