+++ addons/mail/__manifest__.py
@@ -101,9 +101,11 @@
         'static/src/components/moderation_ban_dialog/moderation_ban_dialog.xml',
         'static/src/components/moderation_discard_dialog/moderation_discard_dialog.xml',
         'static/src/components/moderation_reject_dialog/moderation_reject_dialog.xml',
+        'static/src/components/notification_alert/notification_alert.xml',
         'static/src/components/notification_group/notification_group.xml',
         'static/src/components/notification_list/notification_list.xml',
         'static/src/components/notification_popover/notification_popover.xml',
+        'static/src/components/notification_request/notification_request.xml',
         'static/src/components/partner_im_status_icon/partner_im_status_icon.xml',
         'static/src/components/partner_mention_suggestion/partner_mention_suggestion.xml',
         'static/src/components/thread_icon/thread_icon.xml',
