+++ addons/mail/data/mail_data.xml
@@ -1,6 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <data noupdate="1">
+
+        <record id="base.partner_root" model="res.partner">
+            <field name="name">OdooBot</field>
+            <field name="email">odoobot@example.com</field>
+            <field name="image_1920" type="base64" file="mail/static/src/img/odoobot.png"/>
+        </record>
+
         <!-- user root should not receive emails at creation -->
         <record id="base.user_root" model="res.users">
             <field name="notification_type">inbox</field>
