+++ addons/mail/static/src/components/messaging_menu/messaging_menu.xml
