+++ addons/mail/static/src/components/notification_alert/notification_alert.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-    <t t-name="mail_bot.NotificationAlert" owl="1">
+    <t t-name="mail.NotificationAlert" owl="1">
         <div class="o_NotificationAlert">
             <t t-if="env.isMessagingInitialized()">
                 <center t-if="isNotificationBlocked" class="o_notification_alert alert alert-primary">
