+++ addons/mail/static/src/components/notification_request/notification_request.scss
