+++ addons/mail/static/src/components/notification_request/notification_request.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-    <t t-name="mail_bot.NotificationRequest" owl="1">
+    <t t-name="mail.NotificationRequest" owl="1">
         <div class="o_NotificationRequest" t-on-click="_onClick">
             <div class="o_NotificationRequest_sidebar">
                 <div class="o_NotificationRequest_imageContainer o_NotificationRequest_sidebarItem">
