+++ addons/mail/views/assets.xml
@@ -101,9 +101,11 @@
                 <script type="text/javascript" src="/mail/static/src/components/moderation_ban_dialog/moderation_ban_dialog.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/moderation_discard_dialog/moderation_discard_dialog.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/moderation_reject_dialog/moderation_reject_dialog.js"></script>
+                <script type="text/javascript" src="/mail/static/src/components/notification_alert/notification_alert.js"/>
                 <script type="text/javascript" src="/mail/static/src/components/notification_group/notification_group.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/notification_list/notification_list.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/notification_popover/notification_popover.js"></script>
+                <script type="text/javascript" src="/mail/static/src/components/notification_request/notification_request.js"/>
                 <script type="text/javascript" src="/mail/static/src/components/partner_im_status_icon/partner_im_status_icon.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/partner_mention_suggestion/partner_mention_suggestion.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/thread_icon/thread_icon.js"></script>
@@ -160,6 +162,7 @@
                 <script type="text/javascript" src="/mail/static/src/widgets/discuss_invite_partner_dialog/discuss_invite_partner_dialog.js"></script>
                 <script type="text/javascript" src="/mail/static/src/widgets/form_renderer/form_renderer.js"></script>
                 <script type="text/javascript" src="/mail/static/src/widgets/messaging_menu/messaging_menu.js"></script>
+                <script type="text/javascript" src="/mail/static/src/widgets/notification_alert/notification_alert.js"/>
 
                 <!-- notification_list_item.scss contains variables/mixins, it must be positioned before sub-components of NotificationList -->
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/notification_list/notification_list_item.scss"/>
@@ -203,6 +206,7 @@
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/notification_group/notification_group.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/notification_list/notification_list.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/notification_popover/notification_popover.scss"/>
+                <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/notification_request/notification_request.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/partner_im_status_icon/partner_im_status_icon.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/partner_mention_suggestion/partner_mention_suggestion.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/mail/static/src/components/thread_icon/thread_icon.scss"/>
@@ -277,8 +281,8 @@
                 <script type="text/javascript" src="/mail/static/src/components/partner_im_status_icon/partner_im_status_icon_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/partner_mention_suggestion/partner_mention_suggestion_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/thread_icon/thread_icon_tests.js"></script>
-                <script type="text/javascript" src="/mail/static/src/components/thread_preview/thread_preview_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/thread_needaction_preview/thread_needaction_preview_tests.js"></script>
+                <script type="text/javascript" src="/mail/static/src/components/thread_preview/thread_preview_tests.js"/>
                 <script type="text/javascript" src="/mail/static/src/components/thread_textual_typing_status/thread_textual_typing_status_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/components/thread_viewer/thread_viewer_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/models/attachment/attachment_tests.js"></script>
@@ -288,6 +292,7 @@
                 <script type="text/javascript" src="/mail/static/src/utils/throttle/throttle_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/utils/timer/timer_tests.js"></script>
                 <script type="text/javascript" src="/mail/static/src/widgets/form_renderer/form_renderer_tests.js"></script>
+                <script type="text/javascript" src="/mail/static/src/widgets/notification_alert/notification_alert_tests.js"/>
             </xpath>
         </template>
     </data>
