+++ addons/mail/views/res_users_views.xml
@@ -8,11 +8,14 @@
             <field name="model">res.users</field>
             <field name="inherit_id" ref="base.view_users_form_simple_modif"/>
             <field name="arch" type="xml">
-            <data>
-                <field name="email" position="before">
-                    <field name="notification_type" widget="radio" readonly="0"/>
-                </field>
-            </data>
+                <data>
+                    <field name="email" position="before">
+                        <field name="notification_type" widget="radio" readonly="0"/>
+                    </field>
+                    <xpath expr="//field[@name='image_1920']" position="before">
+                        <widget name="notification_alert"/>
+                    </xpath>
+                </data>
             </field>
         </record>
