+++ addons/mail_bot/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     'name': 'OdooBot',
-    'version': '1.1',
+    'version': '1.2',
     'category': 'Productivity/Discuss',
     'summary': 'Add OdooBot in discussions',
     'description': "",
@@ -21,9 +21,5 @@
     ],
     'qweb': [
         'static/src/bugfix/bugfix.xml',
-        'static/src/components/messaging_menu/messaging_menu.xml',
-        'static/src/components/notification_alert/notification_alert.xml',
-        'static/src/components/notification_list/notification_list.xml',
-        'static/src/components/notification_request/notification_request.xml',
     ],
 }
