+++ addons/mail_bot/data/mailbot_data.xml
@@ -1,12 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <data>
-        <record id="base.partner_root" model="res.partner">
-            <field name="name">OdooBot</field>
-            <field name="email">odoobot@example.com</field>
-            <field name="image_1920" type="base64" file="mail/static/src/img/odoobot.png"/>
-        </record>
-
         <record id="base.user_root" model="res.users">
             <field name="odoobot_state">disabled</field>
         </record>
