+++ addons/mail_bot/models/__init__.py
@@ -5,5 +5,4 @@ from . import ir_http
 from . import mail_bot
 from . import mail_channel
 from . import mail_thread
-from . import res_partner
 from . import res_users
