--- addons/mail_bot/static/src/components/messaging_menu/messaging_menu.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates xml:space="preserve">
-    <t t-inherit="mail.MessagingMenu" t-inherit-mode="extension">
-        <xpath expr="//NotificationList" position="attributes">
-           <attribute name="t-on-o-odoobot-request-clicked" t-translation="off">_onOdoobotRequestClicked</attribute>
-        </xpath>
-    </t>
-</templates>
