+++ addons/mail_bot/views/assets.xml
@@ -4,19 +4,11 @@
         <template id="assets_backend" name="mailbot assets" inherit_id="web.assets_backend">
             <xpath expr="script[last()]" position="after">
                 <script type="text/javascript" src="/mail_bot/static/src/bugfix/bugfix.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/components/messaging_menu/messaging_menu.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/components/notification_alert/notification_alert.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/components/notification_list/notification_list.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/components/notification_request/notification_request.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/models/messaging/messaging.js"></script>
                 <script type="text/javascript" src="/mail_bot/static/src/models/messaging_initializer/messaging_initializer.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/models/messaging_menu/messaging_menu.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/widgets/notification_alert/notification_alert.js"></script>
             </xpath>
             <xpath expr="link[last()]" position="after">
                 <link rel="stylesheet" type="text/scss" href="/mail_bot/static/src/scss/odoobot_style.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/mail_bot/static/src/bugfix/bugfix.scss"/>
-                <link rel="stylesheet" type="text/scss" href="/mail_bot/static/src/components/notification_request/notification_request.scss"/>
             </xpath>
         </template>
 
@@ -29,9 +21,7 @@
         <template id="qunit_suite" name="mailbot_tests" inherit_id="web.qunit_suite_tests">
             <xpath expr="." position="inside">
                 <script type="text/javascript" src="/mail_bot/static/src/bugfix/bugfix_tests.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/components/messaging_menu/messaging_menu_tests.js"></script>
                 <script type="text/javascript" src="/mail_bot/static/src/models/messaging_initializer/messaging_initializer_tests.js"></script>
-                <script type="text/javascript" src="/mail_bot/static/src/widgets/notification_alert/notification_alert_tests.js"></script>
             </xpath>
         </template>
     </data>
