+++ addons/mail_bot/views/res_users_views.xml
@@ -27,17 +27,4 @@
             </data>
         </field>
     </record>
-
-    <!-- Blocked notification alert -->
-    <record id="view_users_form_inherit_mail_notification_alert" model="ir.ui.view">
-        <field name="name">res.users.preferences.form.notification.alert</field>
-        <field name="model">res.users</field>
-        <field name="inherit_id" ref="base.view_users_form_simple_modif"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='image_1920']" position="before">
-                <widget name="notification_alert"/>
-            </xpath>
-        </field>
-    </record>
-
 </data></odoo>
